+++ src/Viewer/PickHandler.cpp
@@ -298,7 +298,7 @@ bool PickHandler::handleKeyDown( const osgGA::GUIEventAdapter& ea, GUIActionAdap
 		if ( isCtrlPressed ) {
 			layout->pause();
 			coreGraph->setNodesFreezed( true );
-			layout->getAlg()->setMaxDistance( distance * 0.8 );
+			layout->getAlg()->setMaxDistance( distance * ( float )0.8 );
 			coreGraph->scaleGraphToBase();
 			coreGraph->setNodesFreezed( false );
 			layout->play();
@@ -306,7 +306,7 @@ bool PickHandler::handleKeyDown( const osgGA::GUIEventAdapter& ea, GUIActionAdap
 		else {
 			layout->pause();
 			coreGraph->setNodesFreezed( true );
-			layout->getAlg()->setMaxDistance( distance * 1.2 );
+			layout->getAlg()->setMaxDistance( distance * ( float )1.2 );
 			coreGraph->scaleGraphToBase();
 			coreGraph->setNodesFreezed( false );
 			layout->play();
@@ -431,10 +431,10 @@ bool PickHandler::handleKeyDown( const osgGA::GUIEventAdapter& ea, GUIActionAdap
 		float distance = layout->getAlg()->getMaxDistance();
 
 		if ( isCtrlPressed ) {
-			layout->getAlg()->setMaxDistance( distance * 0.8 );
+			layout->getAlg()->setMaxDistance( distance * ( float )0.8 );
 		}
 		else {
-			layout->getAlg()->setMaxDistance( distance * 1.2 );
+			layout->getAlg()->setMaxDistance( distance * ( float )1.2 );
 		}
 	}
